requirements.txt
aiogram-2.21
environs~=8.0.0